Move chapter picker to components folder
